code functional as is
Notes for Matt:  code is working, but  radio button & feed back text box is not quite working the way we talked about.  If you hit yes (for I like this itinerary) the box goes away.  It works correctly if you hit no. But either way, you see the feedback box when the demo is first opened.  The functionality I didn't get to is adding something that prints once the user likes the itinerary "Great, Have a Fun Trip." Also, I know you talked about adding more functionality to it by providing "was this better than the last?", but I keep getting caught in an error with that.  For formatting, it would be great to have a title for each of our sections.  Probably a no brainer!  I am assuming there's a way to just have the city come up as "Nashville" so this looks a bit more seamless, although you can put in any city and it works.